// UI/src/app/models/setcalendar.ts
export class SetCalender {
    appraisal_term: number;
    appraisal_start_date: string | null;
    objective_set_date: string | null;
    self_appraisal_end_date: string;
    review_end_date: string;
    closer_date: string;
    rating_scale: number;
    qualification_criteria: number; 
    approved_by : number | null
    comments : string | null
 }
  